Move SCC word class into a specific file
